refactor: move createTeamsHandler from trpc to features package
This fixes a circular dependency where packages/features was importing
from packages/trpc. The handler is now located in
packages/features/ee/organizations/lib/createTeams/ and the trpc handler
is now a thin wrapper that imports from features.

The handler now accepts inviteMembersWithNoInviterPermissionCheck as a
function parameter to break the circular dependency.
