+++ packages/features/ee/organizations/lib/createTeams/index.ts
@@ -1,0 +1,1 @@
+export { createTeamsHandler, type CreateTeamsInput } from "./createTeamsHandler";
