+++ packages/features/ee/organizations/lib/service/onboarding/BaseOnboardingService.ts
@@ -24,8 +24,9 @@ import { prisma } from "@calcom/prisma";
 import type { Prisma, Team, User } from "@calcom/prisma/client";
 import { CreationSource, MembershipRole, UserPermissionRole } from "@calcom/prisma/enums";
 import { userMetadata, teamMetadataStrictSchema } from "@calcom/prisma/zod-utils";
-import { createTeamsHandler } from "@calcom/trpc/server/routers/viewer/organizations/createTeams.handler";
 import { inviteMembersWithNoInviterPermissionCheck } from "@calcom/trpc/server/routers/viewer/teams/inviteMember/inviteMember.handler";
+
+import { createTeamsHandler } from "../../createTeams";
 
 import { OrganizationPaymentService } from "../../OrganizationPaymentService";
 import { OrganizationPermissionService } from "../../OrganizationPermissionService";
@@ -534,20 +535,23 @@ export abstract class BaseOnboardingService implements IOrganizationOnboardingSe
       )} teams for organization ${organizationId}`
     );
 
-    await createTeamsHandler({
-      ctx: {
-        user: {
-          ...owner,
-          organizationId,
+    await createTeamsHandler(
+      {
+        ctx: {
+          user: {
+            ...owner,
+            organizationId,
+          },
+        },
+        input: {
+          teamNames: teamsToCreate,
+          orgId: organizationId,
+          moveTeams: teamsToMove,
+          creationSource: CreationSource.WEBAPP,
         },
       },
-      input: {
-        teamNames: teamsToCreate,
-        orgId: organizationId,
-        moveTeams: teamsToMove,
-        creationSource: CreationSource.WEBAPP,
-      },
-    });
+      inviteMembersWithNoInviterPermissionCheck
+    );
 
     log.info(
       `Created ${teamsToCreate.length} teams and moved ${teamsToMove.length} teams for organization ${organizationId}`
